JSCOE-155 count vote
