+++ eventsource-eventbridge-lambda/elections/serverless.yml
@@ -67,6 +67,20 @@ functions:
           retryPolicy:
             maximumEventAge: 86400 # default 1 day
             maximumRetryAttempts: 185 # default 185 attempts
+  countVote:
+    handler: src/handlers/countVote.handle
+    events:
+      - eventBridge:
+          eventBus: ${cf:${self:custom.eventBusStack}-${self:custom.stage}.EventBusArn}
+          pattern:
+            source:
+              - ${self:custom.eventSource}
+            detail-type:
+              - 'Vote Cast'
+          deadLetterQueueArn: !GetAtt CountVoteRuleDLQ.Arn
+          retryPolicy:
+            maximumEventAge: 86400 # default 1 day
+            maximumRetryAttempts: 185 # default 185 attempts
 
 resources:
   Resources:
@@ -79,6 +93,11 @@ resources:
       Type: AWS::SQS::Queue
       Properties:
         QueueName: ${self:service}-${self:custom.stage}-saveVoteRule-DLQ
+
+    CountVoteRuleDLQ:
+      Type: AWS::SQS::Queue
+      Properties:
+        QueueName: ${self:service}-${self:custom.stage}-countVoteRule-DLQ
 
     ElectionsTable:
       Type: 'AWS::DynamoDB::Table'
