+++ eventsource-eventbridge-lambda/elections/src/handlers/countVote.js
@@ -1,21 +1,21 @@
 const ElectionService = require('../services/election');
 
 /**
- * Save vote function request handler. EventBridge event handler.
+ * Count vote function request handler. EventBridge event handler.
  * @param {Object} event The Lambda request event.
  * @param {Object} event.detail The EventBridge `detail`.
  * @param {Object} context TheLambda request context.
  * @returns A Lambda response object.
  */
 exports.handle = async (event, context) => {
-  console.log(`Handler::saveVote`);
+  console.log(`Handler::countVote`);
   console.log(`event:\n${JSON.stringify(event, null, 2)}`);
 
   if (isValidRequest(event)) {
     const { detail = {} } = event;
-    const { ballotId, voter, vote } = detail;
+    const { vote } = detail;
 
-    await ElectionService.saveVote(ballotId, voter, vote);
+    await ElectionService.countVote(vote.electionId, vote.candidateId);
 
     return {
       status: 200,
@@ -36,12 +36,10 @@ exports.handle = async (event, context) => {
  */
 const isValidRequest = (event = {}) => {
   const { detail = {} } = event;
-  const { ballotId, voter = {}, vote = {} } = detail;
+  const { vote = {} } = detail;
 
-  const hasBallotId = !!ballotId && typeof ballotId === 'string';
-  const hasVoterId = !!voter.voterId && typeof voter.voterId === 'string';
   const hasElectionId = !!vote.electionId && typeof vote.electionId === 'string';
   const hasCandidateId = !!vote.candidateId && typeof vote.candidateId === 'string';
 
-  return hasBallotId && hasVoterId && hasElectionId && hasCandidateId;
+  return hasElectionId && hasCandidateId;
 };
