--- eventsource-eventbridge-lambda/electionsummary/README.md
@@ -1,13 +1,0 @@
-# Election Summary Service
-
-This project provisions Lambda function(s), and EventBridge resources to provide _Election Summary_ domain functionality.
-
-This project uses the Serverless Framework.
-
-## Install
-
-## Run
-
-## Remove / Clean Up
-
-## Related Information
