--- eventsource-eventbridge-lambda/electionsummary/serverless.yml
@@ -1,45 +1,0 @@
-service: jscoe155-electionsummary
-
-custom:
-  eventBusStack: 'jscoe155-events'
-  eventSource: 'moduscreate.jscoe155'
-  stage: ${opt:stage, self:provider.stage}
-  tags:
-    App: 'jscoe.moduscreate.com'
-    Env: ${self:custom.stage}
-    OU: 'moduscreate'
-    Owner: 'Matthew Warman'
-    Issue: 'JSCOE-155'
-
-provider:
-  name: aws
-  stage: dev
-  region: us-east-1
-  runtime: nodejs14.x
-  memorySize: 128
-  timeout: 6
-  logRetentionInDays: 7
-  tags: ${self:custom.tags}
-
-functions:
-  voteCast:
-    handler: src/handlers/voteCast.handle
-    events:
-      - eventBridge:
-          eventBus: ${cf:${self:custom.eventBusStack}-${self:custom.stage}.EventBusArn}
-          pattern:
-            source:
-              - ${self:custom.eventSource}
-            detail-type:
-              - 'Vote Cast'
-          deadLetterQueueArn: !GetAtt VoteCastRuleDLQ.Arn
-          retryPolicy:
-            maximumEventAge: 86400 # default 1 day
-            maximumRetryAttempts: 185 # default 185 attempts
-
-resources:
-  Resources:
-    VoteCastRuleDLQ:
-      Type: AWS::SQS::Queue
-      Properties:
-        QueueName: ${self:service}-${self:custom.stage}-voteCastRule-DLQ
